Refactor to polymorphism
